enhance: narrow Marisa CSR element type from uint64_t to uint32_t
Values stored in csr_index_/csr_offsets_ are row offsets within a
single segment. Milvus segments are sized well below 2^32 rows, so a
32-bit slot is plenty and halves both the resident-memory footprint
and the on-disk entry size for these structures.

Guard the assumption with AssertInfo on both the build path
(fill_offsets) and the V3 load path, so a future build that raises
the segment cap trips loudly instead of silently truncating.

+++ internal/core/src/index/StringIndexMarisa.h
@@ -172,10 +172,13 @@ class StringIndexMarisa : public StringIndex {
 
     // CSR (Compressed Sparse Row) format for str_id → offsets mapping.
     // csr_index_[key_id] .. csr_index_[key_id+1] = range in csr_offsets_.
-    std::vector<uint64_t> csr_index_;    // size = num_keys + 1 (memory mode)
-    std::vector<uint64_t> csr_offsets_;  // size = total_num_rows (memory mode)
-    const uint64_t* csr_index_ptr_ = nullptr;    // read accessor (vec or mmap)
-    const uint64_t* csr_offsets_ptr_ = nullptr;  // read accessor (vec or mmap)
+    // Values are row indices within a single segment. uint32_t is safe
+    // because Milvus segments never approach 2^32 rows; fill_offsets and
+    // LoadEntries assert the bound on the build / load paths.
+    std::vector<uint32_t> csr_index_;    // size = num_keys + 1 (memory mode)
+    std::vector<uint32_t> csr_offsets_;  // size = total_num_rows (memory mode)
+    const uint32_t* csr_index_ptr_ = nullptr;    // read accessor (vec or mmap)
+    const uint32_t* csr_offsets_ptr_ = nullptr;  // read accessor (vec or mmap)
     size_t csr_num_keys_ = 0;
     char* csr_mmap_data_ = nullptr;  // mmap region for csr_index + csr_offsets
     int64_t csr_mmap_size_ = 0;
